Generator: add schemas exclusion, merge utility modules
