+++ _generator/overrides.yaml
@@ -7,5 +7,6 @@ excludedSchemaIds:
 types:
   Limitation: '[Limitation](../guidelines/pagination.md#limitation)'
   TimeFilterInterval: '[Time interval](_objects.md#time-interval)'
+  DateTimeIntervalUtc: '[Time interval](_objects.md#time-interval)'
 
 properties: {}
