+++ _generator/utils.js
@@ -4,3 +4,16 @@ import YAML from 'yaml';
 export function loadYaml(path) {
   return YAML.parse(fs.readFileSync(path, 'utf8'));
 }
+/**
+ * @param {string} tagName
+ * @returns {string}
+ */
+export function tagToPageName(tagName) {
+  return tagName.toLowerCase().replace(/[^a-z0-9]/g, '');
+}
+export function slugify(str) {
+  return str.toLowerCase().replace(/[^a-z0-9]/g, '-');
+}
+export function firstLine(str = '') {
+  return str.split(/\r\n|\r|\n/)[0];
+}
